Add new default reference types page
One thing I struggle with as a new Pony rider is figuring out what the
default capability is if one isn't specified.

Judging by https://ponylang.zulipchat.com/#narrow/stream/189985-beginner-help/topic/Default.20reference.20capabilities
I am not the only one.

This new tutorial page was partially informed by the defaults claimed in
the rest of the tutorial, but a lot of it came by way of trial and
compiler error. I tried to come up with as many different "default"
scenarios as I could and implemented them to discover what happened.

Some notes for reviewers:

* I'm not certain I got all of these correct. I have something like two
  hours of Pony experience under my belt. I'm guessing there a massive
  holes and nuances I missed.
* I'm not certain I found all the places default capabilities are used.
* I may not have used the terminology exactly correctly.

That said, I think the layout and format are mostly ok, so this should
be pretty easy to build on.